// src/domain/process/component/process-detail/component/free-field-2/types/free-field-2.type.d.ts
import { BaseType } from '@/core/domain/types/base.type'

export namespace FreeField2Type {
  export type Input = {
    name?: string
  } & BaseType.Input

  export type Output = Input

  export type OutputPaginated = {
    count: number
    limit: number
    offset: number
    data: Partial<Output[]>
  }

  export type Repository = Output

  export type RepositoryPaginated = OutputPaginated
}